+++ src/lang/parser.c
@@ -1,5 +1,6 @@
 #include "posix.h"
 
+#include <inttypes.h>
 #include <limits.h>
 #include <stdarg.h>
 
@@ -199,7 +200,7 @@ node_to_s(struct node *n)
 		i += snprintf(&buf[i], BUF_SIZE_S - i, ":'%s'", n->dat.s);
 		break;
 	case node_number:
-		i += snprintf(&buf[i], BUF_SIZE_S - i, ":%ld", (intmax_t)n->dat.n);
+		i += snprintf(&buf[i], BUF_SIZE_S - i, ":%" PRId64, n->dat.n);
 		break;
 	case node_argument:
 		i += snprintf(&buf[i], BUF_SIZE_S - i, ":%s", n->subtype == arg_kwarg ? "kwarg" : "normal");
